+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-2016.10.04:  
+2016.12.21:
+Changes:
+- new structure
+- immoscout and mydealz working
+
+2016.10.04:
 Changes:
 - os.path.join, cross platform
-- Interface Order
+- Interface order
 - fetch_content fetches multiple Soup objects
 - Dict comprehension to filter out empty searches in json #beproud
 - Started with PriceChecker